Account for non-json responses
Usually CouchDB will always return JSON responses to any API requests.
However I have an Nginx proxy server in front which will sometimes
return a 413 error if the request body is too large. In that case Nginx
won't respond with JSON, and so pouchdb will raise an error when trying
to parse the JSON, and the 413 status code will be lost.