Change winningRev function
There is a bug where calling `bulkDocs` with the `new_edits: true`
option does not work properly with `_deleted: true` documents. See this
issue on github: pouchdb/pouchdb#7562

This bug seems to have to do with how winning revisions are chosen. When
you submit a doc to `bulkDocs` with the `new_edits: false`, it will
compare the revision tree on the submitted doc with the revision trees
on the target doc and choose a winning revision. It does this in the
[`winningRev`](https://github.com/pouchdb/pouchdb/blob/aca46756644b1a4786d41229e7db5d14fce1680c/packages/node_modules/pouchdb-merge/src/winningRev.js#L6)
function, called [here](https://github.com/pouchdb/pouchdb/blob/aca46756644b1a4786d41229e7db5d14fce1680c/packages/node_modules/pouchdb-adapter-utils/src/updateDoc.js#L55).
So if you have a doc like `{_id: "foo" _rev: "6-abc"}` and you submit a
bulkDocs like `[{_id: "foo", _rev: "7-def"}]` you will end up with two
revision trees. One with `[1 2 3 4 5 6-abc]` and one with `[7-def]`. In
this case `7-def` revision will win because it has a higher `pos` (the 7).

This is *not* the case if the `7-def` rev is `deleted: true`. That is
because the `winningRev` algorithim prioritizes (above all else)
non-deleted leaves of a branch. So because the `[1 2 3 4 5 6-abc]`
branch ends in a non-deleted leaf, it will always override the `[7-def]`
branch even though 7 is higher than 6.

Normally when pouchdb pulls data from another instance, it will include
the entire revision history, so the revision tree from the local doc
(`[1 2 3 4 5 6-abc]`) will get merged with the remote one
(`[1 2 3 4 5 6-foo 7-def]`), and the `7-def` revision will win even
though it is deleted because it is the the leaf of the only branch (so
there are no other leaves to override it).

If we remove the prioritization of non-deleted leaves, then we recover
the functionality of deleting with `bulkDocs` with `new_edits: true`,
without having to passing in the entire revision history.

The decision for which revision to choose is supposed to be arbitrary,
but deterministic, so changing this might have unintended side effects.

> By default, CouchDB picks one arbitrary revision as the “winner”,
> using a deterministic algorithm so that the same choice will be made on
> all peers. The same happens with views: the deterministically-chosen
> winner is the only revision fed into your map function.
> – https://docs.couchdb.org/en/stable/replication/conflicts.html

The implementation in PouchDB seems to be based on a sidebar from this
page: https://web.archive.org/web/20160625121023/http://guide.couchdb.org/draft/conflicts.html

> ### How Does CouchDB Decide Which Revision to Use?
>
> CouchDB guarantees that each instance that sees the same conflict
> comes up with the same winning and losing revisions. It does so by
> running a deterministic algorithm to pick the winner. The application
> should not rely on the details of this algorithm and must always
> resolve conflicts. We’ll tell you how it works anyway.
>
> Each revision includes a list of previous revisions. The revision with
> the longest revision history list becomes the winning revision. If
> they are the same, the _rev values are compared in ASCII sort order,
> and the highest wins. So, in our example,
> 2-de0ea16f8621cbac506d23a0fbbde08a beats 2-7c971bb974251ae8541b8fe045964219.
>
> One advantage of this algorithm is that CouchDB nodes do not have to
> talk to each other to agree on winning revisions. We already learned
> that the network is prone to errors and avoiding it for conflict
> resolution makes CouchDB very robust.